Przeniesienie pliku postac.h do klasy.h
