+++ klasy.h
@@ -1,9 +1,16 @@
 #include<iostream>
-#include "postac.h"
 #include <string>
 
 using namespace std;
 
+class Postac {
+    public:
+        virtual void atakuj(Postac &cel) = 0;
+        virtual void przyjmij_atak(int) = 0;
+        virtual void wypisz_statystyki() = 0;
+        virtual bool czy_zyje() = 0;
+        virtual void awansuj() = 0;
+};
 class Strzelec : public Postac {
     public:
         Strzelec();
